Disable PR builds for public repos
As things stand right now, PR builds for public repositories are unsafe,
due to the fact that anyone can submit a PR that modifies Jenkinsfile.

As far as I am concerned, this is a regression; it used to be that this
plugin did not automatically pick up PR builds.

There are various efforts in progress to address this in different ways,
and there are also other project recognizers that are considered that
can change the situation.

But until that happens, please disable this feature, for it's too easy
for people to "incorrectly" use this functionality with public
repositories and make their Jenkins instances vulnerable.