More efficient job dispatching in PR GHEventSubscriber
Upon receiving a webhook for a multibranch project, we follow the same
sort of process that branch indexing does: decide if there is a PR build
that needs to happen, and schedule a build with the right name and
revision.

I'm not taking note of branch criteria here, nor even whether the jobs
already exist. I'm not sure if that matters. Let's find out! :/