fix(discover branches): keep "PR branches" option
This PR is in response to the branch-discovery help text:
"
Only branches that are also filed as PRs
This option exists to preserve legacy behaviour
when upgrading from older versions of the plugin.
NOTE: If you have an actual use case for this option
please file a pull request against this text.
"

We have a Jenkins pipeline that deploys dockers for
pull requests to a k8s cluster.  The Dockers are
built by a separate process, and tagged with the
branch name, so we want our k8s-deploy pipeline
to be fed the PR branch name in env.GIT_BRANCH,
so it can reference the appropriate docker.

Hope that makes sense.