Fix identity comparison for GitHubSCMSource
When scanning for new runs, PullRequestSCMRevision instances are compared.  While technically one combination of
base and pull commit produce the same merge, if we detect a new merge commit we should still treat that as a new revision.

For example:
Before this change, if a user closes a PR (unmerged) and then reopenes it, a repository scan would not treat the PR as changed.

After this change, a closing, reopening, and then scan will result in change being detected with a message like this:

Changes detected: PR-3-merge (
58caa961c3fc5d68268e0d860670041b97cc86b0+edb61773c0185852a5d5c7afc17a924fee730363 (7c907cc0d0eaf15f2b84d030e23f5fd96e1b59d4) ->
58caa961c3fc5d68268e0d860670041b97cc86b0+edb61773c0185852a5d5c7afc17a924fee730363 (e0dd5f238a71951e6a48dc01443fb624a44ff278))