[JENKINS-57257] FileSytem for PRs should always load from target repo
All GitHub PRs can get the commits and contents from the target repo.

Not sure why they used to pull from the branchSource but there is no reason for that.

Without this change PRs from forks simply don't work. This bug made it into the release
due to complexity of the testing framework making it easy to test for incorrect values.