JENKINS-56996 - PRs where parts are not found are orphaned
Bad PRs can happen a number of ways when constituent parts of
the PR were missing, source fork, user, target branch, etc.
The generally mean the PR is permanenently dead.

However, before this change one bad PR would throw an exception
and crash an entire scan. In the case of GitHub Org Folder Pipelines
this would happen all too easily and would stop the whole org from scanning.

This change treats the FileNotFound exceptions thrown by these bad PRs
as errors local to the PR, treats them as orphaned, and continues the scan.

This change only effects PRs, not tags or branches, which are
less prone to errors (or are already doing a similiar thing).