JENKINS-57583 Ignore merge commmit when comparing revisions
Other plugins assume that there are are only two revisions
involved in PR revision identity. Since the mergeHash is generated
based on the base and source hashes, it is not particularly
interesting to use it and can cause bad behavior if we do.

The only place where the mergeHash is interesting is when
getting Jenkinsfile or doing readTrusted. It is an optimization
that should have as few side effects as possible.