JENKINS-60730 add new trait to filter out Branches and Tags older than X
A new trait called BranchOrTagAgeDiscoveryTrait is introduced to filter out branches and tags that are older than `maxDaysOld`. The `maxDaysOld` parameter can be set in the UI when the trait is selected. All branches and tags older than the specified amount of days are ignored after that.

It should not be possible to set `maxDaysOld` to `0`, but if it is `0`, negative or if the returned `lastModified` date of a commit for any reason (incompatibility, error or anything), the branch, tag is not excluded and default behaviour takes over, basically deactivating the trait filter.