Fix windows test
In previous versions of github-api, GHContent.read() (called by contentAsString())
would pull from the "raw" url of the GHContent instance.
Thus on windows, if somebody did not configure Git correctly,
the checkout may have "fixed" line endings that we needed to handle.
The problem with the raw url data is that it can get out of sync when from the actual content.
The GitHub API info stays sync'd and correct, so now GHContent.read() pulls from mime encoded data