Refactor Connector for better GitHubAppCredentials behavior
GitHubAppCredentials could not use the existing Connector.connect() and so created its own
GithubBuilder.  However, that means that is missing a number of standard settings provided by
Connector.connect(), including okhttp, Jenkins proxy settings, and rate limit handling.

This change refactors Connector to add an internal createGitHubBuilder() method that returns a
GitHubBuilder with those features configured. For simplicity, The returned GitHubBuilder
is not cached and also does not cache responses.  If there turns out to be a need for it,
that behavior can be added later.