Fix connection caching multiple release issue
The calls to connect and release for GitHub connections were not matched up.
We ended up having a few code paths with multiple release calls, resulting in errors.

This change fixes the mismatches and makes it less likely that future changes will break this.
it also makes the cache cleaning less agressive, so that if this does occur again connections will be
less likely to get closed while still in use.