Remove rechecking/refreshed rate limit functionality
The ApiRateLimitChecker included functionality to handle when the rate limit is reset "before expected".
If a real world scenario exists for this behavior, it is an extreme edge case. Meanwhile, the functionality
add unneeded complexity to the checkers and is not supported by the github-api RateLimitChecker class.

This change remove this extraneous functionality and fixes the test match.