Start using Google Format auto formatting
This change turns on auto formatting using the Google Format for Java tool.

This change runs checks during the build and will fail the build during the verify phase if the
formatting checks fail.  The failure message tells users to run `spotless:apply` to fix formatting.
After running `spotless:apply` the formatting will be correct.

This is the squashed commit for #392 https://github.com/jenkinsci/github-branch-source-plugin/pull/392 .
Due to my clicking the wrong button, I did a merge rather than a squash in the PR.
This change causes a "diff wall", so squashing is the better option to make the depth of that wall only one commit.