Prevent Potential NPE.
NPE observed was

```
java.lang.NullPointerException
  at org.jenkinsci.plugins.github_branch_source.GitHubSCMNavigator.visitSources(GitHubSCMNavigator.java:995)
  at jenkins.branch.OrganizationFolder.computeChildren(OrganizationFolder.java:535)
  at com.cloudbees.hudson.plugins.folder.computed.ComputedFolder.updateChildren(ComputedFolder.java:278)
  at com.cloudbees.hudson.plugins.folder.computed.FolderComputation.run(FolderComputation.java:166)
  at jenkins.branch.OrganizationFolder$OrganizationScan.run(OrganizationFolder.java:917)
  at hudson.model.ResourceController.execute(ResourceController.java:99)
  at hudson.model.Executor.run(Executor.java:432)
```

Whilst I am not sure of the version that this happened in (it was either 2.11.4 or 2.11.3).  This would be the fix if it was 2.11.4 and is safe either way - we already know repoOwner is not null due to the check at the start of the method for `repoOwner.isEmpty` and the GH API for `getOwnerName` is a bit funky and does point to some GitHub strangeness. https://github.com/hub4j/github-api/blob/a3a5a32e8d9447a46675d1439ba75e1e5a0ecf14/src/main/java/org/kohsuke/github/GHRepository.java#L599-L604

if the issue actually occured in the older version at this line - then
the TaskListener.getLogger() is returning `null` despite the fact that
the method is annotated `NonNull` and the TaskListener
(StreamTaskListener) getLogger() is looking safe on first inspection.