[JENKINS-60874] Configuring ForkPullRequestDiscoveryTrait using Job DSL
JENKINS-26535 describes an issue where wildcards in generics aren't
handled well and result in the inability to configure classes that use
wildcards in Job DSL, such as the SCMHeadAuthority class used for the
forked PR trust policy. To work around this issue, create a new abstract
class named GitHubForkTrustPolicy which extends the necessary
SCMHeadAuthority superclass with its proper generics and use it in the
constructor instead. Update all the trust policies to extend from this
abstract class so they can be used.

Fixes JENKINS-60874.