Prevent testProviderRefresh test failures due to log ordering differences
Accept either the typical ordering of log messages or an alternate
ordering of log messages in testProviderRefresh.  The same messages are
used in each case, but the sequence of messages changes.

Accepting either ordering assures that we won't have plugin BOM failures
due to timing differences or other differences in test infrastructure.

https://github.com/jenkinsci/bom/pull/5718#issuecomment-3324993588
detected a failure in this test when run within the plugin bill of
materials.

https://github.com/jenkinsci/bom/pull/5720#issuecomment-3328261892 tested
more deeply and found that it failed more than 5% of my test runs locally.

Additional local testing confirmed that the order of the log messages was
sometimes slightly different than the original assertion expected. This
change accepts either ordering.

Testing done:

* Before this change, the test would fail 5% of the time on my
  AMD Ryzen 5 5600X 6-Core Processor on Red Hat Enterprise Linux 8
  when run with
  `mvn -Dtest=GithubAppCredentialsTest#testProviderRefresh test`

* After this change, the test passes 100% of the time in the 36 test
  runs that I used to check it